requirements.txt
# Dependencies
pandas>=2.0
numpy>=1.24
pyarrow>=14.0
requests>=2.31
python-dotenv>=1.0
rapidfuzz>=3.5
fuzzywuzzy>=0.18
python-Levenshtein>=0.21
linearmodels>=5.3
statsmodels>=0.14
tabulate>=0.9

# Optional: AI Integration
openai>=1.0
anthropic>=0.20
google-generativeai>=0.5
